fix: properly handle brackets in evaluated expressions
Before this fix, `{` is the expression prefix of the EL engine.
As this character can be in regular strings, out of any expression language, they are replaced with `{'{'}` which is a interpreted as a list of one `{` character during expression evaluation.
That causes problems when `{` is, as a string, inside a expression language function parameter, for example `{#request.pathInfo.matches('/[A-Z]{2}')}` : the `{2}` shouldn't be escaped.

With this fix, `{#` is the expression prefix of the EL engine.
EL prefixes from user input : `{#`, `{(`, or `{T` are replaced by `{##`, `{#(`, or `{#T`.
Regular `{` characters aren't interpreted as expression prefixes by EL SpelExpressionParser