feat: add evalNow to evaluate EL
Since the beginning of the v4, we’ve made some evolutions on the
template engine to support evaluation in a reactive way. At the time, we
decided to deprecate blocking getValue() and convert() methods and force
using the reactive eval() method instead.

We now think that non-reactive evaluation of EL expression can still
make sense when resolving an EL expression outside of the context of an
API call:

It offers better performances when evaluating simple EL expressions.

It allows us to keep simplicity. Especially when you need to evaluate an
EL outside of a reactive chain.